Extracted static city loading from CityService into StaticCityLoader, so we can use it for bazaar terminals later
