+++ src/main/java/com/projectswg/holocore/resources/support/data/server_info/loader/ServerData.kt
@@ -105,6 +105,7 @@ object ServerData {
 	val destroyMissions		by SoftDataLoaderDelegate(::DestroyMissionLoader)
 	val dynamicLairs		by SoftDataLoaderDelegate(::DynamicLairLoader)
 	val rawResources		by SoftDataLoaderDelegate(::RawResourceLoader)
+	val staticCities		by SoftDataLoaderDelegate(::StaticCityLoader)
 
 	private class WeakDataLoaderDelegate<T: DataLoader>(loaderCreator: () -> T): DataLoaderDelegate<T>(::WeakReference, loaderCreator)
 	private class SoftDataLoaderDelegate<T: DataLoader>(loaderCreator: () -> T): DataLoaderDelegate<T>(::SoftReference, loaderCreator)
